Fix the `require_dependency` main usage for STI workaround:
- One the main usage of the `require_dependency` inside an application
  is to workaround the STI issue described in this Rails [documentation](https://guides.rubyonrails.org/autoloading_and_reloading_constants.html#autoloading-and-sti)

  This commit aims to remove this requirement by loading all the
  STI types from the database instead of the descendants.

- Zeitwerk claims that "All known use cases of require_dependency have been eliminated. you should grep the project and delete them".
  I'm not sure whether I correctly understand this affirmation,
  but removing all `require_dependency` call from an application using
  Zeitwerk doesn't solve the problem unless we do some [preloading](https://github.com/fxn/zeitwerk#preloading)
