+++ activerecord/test/models/autoloadable/necromancer.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Necromancer < Thaumaturgist
+end
